The first Machine Learning assignment
An implementation of the Nearest Neighbor methods for classification.

The datasets used were the built-in iris dataset from scikit-learn module of python, and the ionosphere dataset.

Nearest Neighbor method ran on both datasets we calculate the error rate it makes on our test set.

Finally we implement a conformal predictor based on the nearest neighbor conformity measure:

distance to nearest sample of a different class / distance to nearest sample of the same class