second machine learning assignment
this assignment contains an inductive conformal predictor for regression.

the datasets used were two variants of the diabetes dataset, the scikit-learn built-in version and the original dataset which can be found in 'https://hastie.su.domains/Papers/LARS/diabetes.data'

the training and test coefficient of determination is found for the Lasso model for both datasets.

they are then preprocessed using StandardScaler function while avoiding data snooping.

finally we implement an inductive conformal predictor, and found prediction interval for each test sample for significance levels 5% and 20%.