the last machine learning assignment
there are two datasets used in this assignment, the wine dataset which is buil into the scikit-learn module, and the USPS dataset of handwritten digits. (the training and test data can be found at the bottom ogf the page) https://hastie.su.domains/ElemStatLearn/

we estimate the generalisation-accuracy of the support vector machine (SVM) using cross validation on the training sets.

Additionally the error rate of the SVM is found using the default values of parameters of the SVM.